refactor: refactor page routes
