--- src/pages/component/uploadmusic/uploadmusic.vue
@@ -1,48 +1,0 @@
-<template>
-  <view>
-    <!-- pages/component/uploadmusic/uploadmusic.wxml -->
-    <text>pages/component/uploadmusic/uploadmusic.wxml</text>
-  </view>
-</template>
-
-<script>
-// pages/component/uploadmusic/uploadmusic.js
-export default {
-  data() {
-    return {};
-  }, /**
-     * 生命周期函数--监听页面加载
-     */
-  onLoad(options) {},
-  /**
-     * 生命周期函数--监听页面初次渲染完成
-     */
-  onReady() {},
-  /**
-     * 生命周期函数--监听页面显示
-     */
-  onShow() {},
-  /**
-     * 生命周期函数--监听页面隐藏
-     */
-  onHide() {},
-  /**
-     * 生命周期函数--监听页面卸载
-     */
-  onUnload() {},
-  /**
-     * 页面相关事件处理函数--监听用户下拉动作
-     */
-  onPullDownRefresh() {},
-  /**
-     * 页面上拉触底事件的处理函数
-     */
-  onReachBottom() {},
-  /**
-     * 用户点击右上角分享
-     */
-  onShareAppMessage() {},
-  methods: {},
-};
-</script>
-<style></style>
