+++ src/pages/quizzes/details.vue
