+++ src/pages/quizzes/random.vue
