+++ src/pages/tools/dictionary.vue
