+++ src/pages/words/characters/details.vue
