+++ src/pages/words/pronunciations/ranking.vue
