+++ src/pages/words/pronunciations/upload.vue
